refactor: move python code into sub folders - use python 3.8 - setup properly with poetry - refactor test imports - setup for using Snowpark
