+++ code/python/normalize_proc_names/normalize_proc_names/__init__.py
@@ -1,0 +1,1 @@
+from normalize_proc_names.normalize_proc_names import *
