+++ code/python/normalize_proc_names/pyproject.toml
@@ -3,13 +3,16 @@ name = "snowflake-playgrounds"
 version = "0.1.0"
 description = ""
 authors = ["James Goodhand <james.goodhand@hey.com>"]
+packages = [{include = "normalize_proc_names"}]
 
 [tool.poetry.dependencies]
-python = "^3.8"
-pytest = "^7.2.1"
-snapshottest = "^0.6.0"
+python = "3.8.*"
 
 [tool.poetry.dev-dependencies]
+
+[tool.poetry.group.test.dependencies]
+pytest = "^7.2.1"
+snapshottest = "^0.6.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
