+++ code/python/tidy_playground/.python-version
@@ -1,0 +1,1 @@
+3.8
