+++ code/python/tidy_playground/pyproject.toml
@@ -1,0 +1,18 @@
+[tool.poetry]
+name = "tidy-playground"
+version = "0.1.0"
+description = ""
+authors = ["James Goodhand <james.goodhand@hey.com>"]
+readme = "README.md"
+packages = [{include = "tidy_playground"}]
+
+[tool.poetry.dependencies]
+python = "3.8.*"
+snowflake-snowpark-python = {extras = ["pandas"], version = "^1.1.0"}
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^7.2.1"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
